% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_glassdoor_pid.R
\name{gd_pid}
\alias{gd_pid}
\alias{gd_pat}
\alias{have_gd_pid}
\alias{have_gd_pat}
\alias{have_gd_tokens}
\title{Glassdoor Partner ID}
\usage{
gd_pid(token = NULL, error = TRUE)

gd_pat(token = NULL, error = TRUE)

have_gd_pid(token = NULL)

have_gd_pat(token = NULL)

have_gd_tokens()
}
\arguments{
\item{token}{Partner ID or Authentication token}

\item{error}{Should the function error if no token specified?}
}
\value{
A vector of class character
}
\description{
Glassdoor Partner ID
}
\examples{
if (have_gd_pid()) {
gd_pid()
}
}
