% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_review.R
\name{gd_review}
\alias{gd_review}
\title{Glassdoor Company Search}
\usage{
gd_review(employer_id, ..., page_number = NULL, page_size = NULL, query = NULL)
}
\arguments{
\item{employer_id}{Glassdoor ID for the company, can be accessed using
\code{\link{gd_company}}}

\item{...}{arguments to pass to \code{\link{gd_api}}}

\item{page_number}{Page number to retrieve. Default is 1.}

\item{page_size}{Page size, i.e. the number of jobs returned on
each page of results. Default is 20.}

\item{query}{Additional options to pass to the query other than those
specified here}
}
\description{
Glassdoor Company Search
}
\examples{
if (have_gd_tokens()) {
res = gd_review(employer_id = 715) # walmart
}

}
