% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip.R
\name{get_ip}
\alias{get_ip}
\title{Get IP Address}
\usage{
get_ip(agent = gd_user_agent())
}
\arguments{
\item{agent}{User Agent for Header}
}
\value{
Character Vector of IP
}
\description{
Get IP Address
}
\note{
Inspired by
\url{https://github.com/gregce/ipify/blob/master/R/ipify.R}
}
\examples{
get_ip()
}
