% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassdoor_job_prog.R
\name{glassdoor_job_prog}
\alias{glassdoor_job_prog}
\title{Glassdoor Job Progression}
\usage{
glassdoor_job_prog(..., jobTitle, countryId = 1, query = NULL)
}
\arguments{
\item{...}{arguments to pass to \code{\link{glassdoor_api}}}

\item{jobTitle}{Job Title - the job title to get related jobs for}

\item{countryId}{Country Id - only 1 (US) is supported right now.}

\item{query}{Additional options to pass to the query other than those
specified here}
}
\description{
Glassdoor Job Progression
}
\examples{
if (have_glassdoor_tokens()) {

paste0("&countryId=1&jobTitle=cashier")
res = glassdoor_job_prog(jobTitle = "cashier", countryId = 1)
}

}
