% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_glassdoor_pid.R
\name{glassdoor_pid}
\alias{glassdoor_pid}
\alias{glassdoor_pat}
\alias{have_glassdoor_pid}
\alias{have_glassdoor_pat}
\title{Glassdoor Partner ID}
\usage{
glassdoor_pid(token = NULL, error = TRUE)

glassdoor_pat(token = NULL, error = TRUE)

have_glassdoor_pid(token = NULL)

have_glassdoor_pat(token = NULL)
}
\arguments{
\item{token}{Partner ID or Authentication token}

\item{error}{Should the function error if no token specified?}
}
\value{
A vector of class character
}
\description{
Glassdoor Partner ID
}
\examples{
if (have_glassdoor_pid()) {
glassdoor_pid()
}
}
