% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassdoor_user_agent.R
\name{glassdoor_user_agent}
\alias{glassdoor_user_agent}
\title{Glassdoor User Agent}
\usage{
glassdoor_user_agent()
}
\value{
Character vector
}
\description{
Glassdoor User Agent
}
\examples{
glassdoor_user_agent()

}
