% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\alias{results.default}
\alias{results.gd_api}
\alias{results.gd_job_prog}
\title{Results of a Glassdoor object}
\usage{
results(gd, ...)

\method{results}{default}(gd, ...)

\method{results}{gd_api}(gd, ...)

\method{results}{gd_job_prog}(gd, ...)
}
\arguments{
\item{gd}{an object for which we want the results, the output
from a Glassdoor call}

\item{\dots}{Any additional arguments to be passed to \code{results}.}
}
\value{
A class of \code{gd_job_prog} will return a
list of \code{progression_table} and \code{job_info}.
}
\description{
Calculates the results from Glassdoor object
}
\examples{
if (have_gd_tokens()) {
   gd = gd_company("walmart")
   res = results(gd)
}
}
